requirements.txt
# Dépendances principales de l'application
streamlit>=1.33,<2
pandas>=2.2
numpy>=1.26
plotly>=5.22
yfinance>=0.2.36
python-dateutil>=2.8.2
types-python-dateutil>=2.8.3
openai>=1.33
pycountry>=22.3.5
scipy>=1.11
